+++ Plutarch/TermCont.hs
@@ -17,8 +17,7 @@ module Plutarch.TermCont (
 ) where
 
 import Data.Kind (Type)
-import Plutarch.Bool (pif)
-import Plutarch.Builtin.Bool (PBool)
+import Plutarch.Builtin.Bool (PBool, pif)
 import Plutarch.DataRepr (HRec, PDataFields, PFields, pletFields)
 import Plutarch.DataRepr.Internal.Field (
   BindFields,